requirements.txt
# Discord RAG Bot - Backend Dependencies

# Core ML Libraries
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3

# LangChain for document processing
langchain==0.1.0
langchain-community==0.0.20

# LLM Integration
openai==1.12.0
requests==2.31.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Environment and utilities
python-dotenv==1.0.0
tiktoken==0.5.2

# PDF processing
pypdf==3.17.4

# Logging
loguru==0.7.2

# Optional: MongoDB (for future upgrade)
pymongo==4.6.1
